Convert UUIDs into actual UUIDs instead of storing as String. (#86)
UUID.fromString is not an inexpensive operation. It's not hugely bad,
but it was one of the things that got highlighted while trying to track
down the source of lag on my multiplayer server, especially where it was
happening with every tick (Modifier Mundane).

I've tested the code to the extend that it compiles and launches and I
haven't crashed after jetting around a test world for a few minutes.

I don't believe this causes any other issues.